+++ cmd/client/export.go
@@ -2,26 +2,22 @@ package client
 
 import (
 	"bufio"
-	"crypto/ecdsa"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"strings"
 
-	"github.com/google/uuid"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	"github.com/ethereum/go-ethereum/accounts/keystore"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
 
+	"github.com/okex/exchain/app/crypto/ethkeystore"
+	"github.com/okex/exchain/app/crypto/hd"
 	"github.com/okex/exchain/libs/cosmos-sdk/client/flags"
 	"github.com/okex/exchain/libs/cosmos-sdk/client/input"
 	"github.com/okex/exchain/libs/cosmos-sdk/crypto/keys"
 	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
-	"github.com/okex/exchain/app/crypto/ethsecp256k1"
-	"github.com/okex/exchain/app/crypto/hd"
 )
 
 // UnsafeExportEthKeyCommand exports a key with the given name as a private key in hex format.
@@ -68,14 +64,14 @@ func UnsafeExportEthKeyCommand() *cobra.Command {
 				return err
 			}
 
-			// Converts key to Ethermint secp256 implementation
-			emintKey, ok := privKey.(ethsecp256k1.PrivKey)
-			if !ok {
+			// Converts tendermint  key to ethereum key
+			ethKey, err := ethkeystore.EncodeTmKeyToEthKey(privKey)
+			if err != nil {
 				return fmt.Errorf("invalid private key type, must be Ethereum key: %T", privKey)
 			}
 
 			// Formats key for output
-			privB := ethcrypto.FromECDSA(emintKey.ToECDSA())
+			privB := ethcrypto.FromECDSA(ethKey)
 			keyS := strings.ToLower(hexutil.Encode(privB)[2:])
 
 			fmt.Println(keyS)
@@ -143,16 +139,14 @@ func ExportEthCompCommand() *cobra.Command {
 				return err
 			}
 
-			// Get eth keystore key by Name
-			ethKey, err := getEthKeyByName(kb, accountName, decryptPassword)
+			// exports private key from keybase using password
+			privKey, err := kb.ExportPrivateKeyObject(accountName, decryptPassword)
 			if err != nil {
 				return err
 			}
 
-			// Export Key to keystore file
-			if err := exportKeyStoreFile(ethKey, encryptPassword, fileName); err != nil {
-				return err
-			}
+			// Exports private key from keybase using password
+			ethkeystore.CreateKeystoreByTmKey(privKey, fileName, encryptPassword)
 
 			fmt.Printf("The keystore has exported to: %s \n", fileName)
 			return nil
@@ -170,56 +164,4 @@ func pathExist(path string) bool {
 		return false
 	}
 	return true
-}
-
-// exportKeyStoreFile Export Key to  keystore file
-func exportKeyStoreFile(ethKey *keystore.Key, encryptPassword, fileName string) error {
-	// Encrypt Key to get keystore file
-	content, err := keystore.EncryptKey(ethKey, encryptPassword, keystore.StandardScryptN, keystore.StandardScryptP)
-	if err != nil {
-		return fmt.Errorf("failed to encrypt key: %s", err.Error())
-	}
-
-	// Write to keystore file
-	err = ioutil.WriteFile(fileName, content, os.ModePerm)
-	if err != nil {
-		return fmt.Errorf("failed to write keystore: %s", err.Error())
-	}
-	return nil
-}
-
-// getEthKeyByName Get eth keystore key by Name
-func getEthKeyByName(kb keys.Keybase, accountName, decryptPassword string) (*keystore.Key, error) {
-	// Exports private key from keybase using password
-	privKey, err := kb.ExportPrivateKeyObject(accountName, decryptPassword)
-	if err != nil {
-		return nil, err
-	}
-
-	// Converts key to Ethermint secp256 implementation
-	emintKey, ok := privKey.(ethsecp256k1.PrivKey)
-	if !ok {
-		return nil, fmt.Errorf("invalid private key type, must be Ethereum key: %T", privKey)
-	}
-
-	//  Converts Ethermint secp256 implementation key to keystore key
-	ethKey, err := newEthKeyFromECDSA(emintKey.ToECDSA())
-	if err != nil {
-		return nil, fmt.Errorf("failed convert to ethKey: %s", err.Error())
-	}
-	return ethKey, nil
-}
-
-// newEthKeyFromECDSA new eth.keystore Key
-func newEthKeyFromECDSA(privateKeyECDSA *ecdsa.PrivateKey) (*keystore.Key, error) {
-	id, err := uuid.NewRandom()
-	if err != nil {
-		return nil, fmt.Errorf("Could not create random uuid: %v", err)
-	}
-	key := &keystore.Key{
-		Id:         id,
-		Address:    ethcrypto.PubkeyToAddress(privateKeyECDSA.PublicKey),
-		PrivateKey: privateKeyECDSA,
-	}
-	return key, nil
 }
